requirements.txt
# --- Web Framework ---
fastapi==0.110.1
uvicorn[standard]==0.29.0

# --- Data Modeling & Validation ---
pydantic==2.6.4
typing-extensions==4.12.2  # Upgraded to avoid conflicts

# --- Async HTTP Client ---
httpx==0.27.0
aiofiles==23.2.1

# --- Core LLM Support (vLLM-compatible + Hugging Face + Bedrock) ---
openai==1.14.3                  # Used for vLLM-compatible APIs
huggingface_hub==0.22.2         # Hosted or local Hugging Face models
boto3==1.34.74                   # AWS Bedrock support
botocore==1.34.74

# --- Caching & Redis Support ---
redis==5.0.1

# --- Data & Metrics ---
pandas==2.2.2
pyarrow==15.0.2

# --- Security ---
python-jose[cryptography]==3.3.0

# --- Dependency Injection (Optional) ---
injector==0.21.0

# --- Testing & Dev Tools ---
pytest==8.2.0
pytest-asyncio==0.23.6
black==24.4.2
isort==5.13.2


# Caching
redis>=5.0.0

python-dotenv